Arreglos vidas

+++ Assets/Nivel 1/Scripts_Level1/VIdas/VidaPlayer_3.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class VidaPlayer_3 : MonoBehaviour
 {
-    public int vida = 100;
+    public float vida = 100;
     public int vidaMaxima = 100;
     public TextMeshProUGUI txtVida;
